Fixing Ping()
In the original code Ping() is expected to return nil. However, in
RecvReply it doesn't check for errors *after* the connection tis
successful. Well, technically it did, but returned nil instead of the
error. So even on a non-authenticated connection ping would *never*
return the err informing you you needed to authenticate first.
I've fixed this now.